# Improve optimization time for partitioned tables in Orca