--- a/src/backend/gporca/libgpos/include/gpos/string/CWStringConst.h
+++ b/src/backend/gporca/libgpos/include/gpos/string/CWStringConst.h
@@ -46,6 +46,16 @@ public:
 
 	// returns the wide character buffer storing the string
 	const WCHAR *GetBuffer() const override;
+
+	// equality
+	static BOOL Equals(const CWStringConst *string1,
+					   const CWStringConst *string2);
+
+	// hash function
+	static ULONG HashValue(const CWStringConst *string);
+
+	// checks whether the string is byte-wise equal to another string
+	BOOL Equals(const CWStringBase *str) const override;
 };
 }  // namespace gpos