--- a/src/backend/gporca/libgpos/src/string/CWStringConst.cpp
+++ b/src/backend/gporca/libgpos/src/string/CWStringConst.cpp
@@ -13,6 +13,7 @@
 
 #include "gpos/base.h"
 #include "gpos/common/clibwrapper.h"
+#include "gpos/utils.h"
 
 using namespace gpos;
 
@@ -118,4 +119,31 @@ CWStringConst::GetBuffer() const
 	return m_w_str_buffer;
 }
 
+// equality
+BOOL
+CWStringConst::Equals(const CWStringConst *string1,
+					  const CWStringConst *string2)
+{
+	ULONG length = GPOS_WSZ_LENGTH(string1->GetBuffer());
+	return length == GPOS_WSZ_LENGTH(string2->GetBuffer()) &&
+		   0 == clib::Wcsncmp(string1->GetBuffer(), string2->GetBuffer(),
+							  length);
+}
+
+// hash function
+ULONG
+CWStringConst::HashValue(const CWStringConst *string)
+{
+	return gpos::HashByteArray(
+		(BYTE *) string->GetBuffer(),
+		GPOS_WSZ_LENGTH(string->GetBuffer()) * GPOS_SIZEOF(WCHAR));
+}
+
+// checks whether the string is byte-wise equal to another string
+BOOL
+CWStringConst::Equals(const CWStringBase *str) const
+{
+	GPOS_ASSERT(nullptr != str);
+	return CWStringBase::Equals(str->GetBuffer());
+}
 // EOF