--- a/src/backend/gporca/libnaucrates/include/naucrates/md/CDXLExtStats.h
+++ b/src/backend/gporca/libnaucrates/include/naucrates/md/CDXLExtStats.h
@@ -52,7 +52,7 @@ private:
 	CMDName *m_mdname;
 
 	// DXL string for object
-	CWStringDynamic *m_dxl_str;
+	CWStringDynamic *m_dxl_str = nullptr;
 
 	CMDDependencyArray *m_dependency_array;
 
@@ -74,7 +74,7 @@ public:
 	CMDName Mdname() const override;
 
 	// DXL string representation of cache object
-	const CWStringDynamic *GetStrRepr() const override;
+	const CWStringDynamic *GetStrRepr() override;
 
 	// serialize relation stats in DXL format given a serializer object
 	void Serialize(gpdxl::CXMLSerializer *) const override;