--- a/src/backend/gporca/libnaucrates/include/naucrates/md/CMDFunctionGPDB.h
+++ b/src/backend/gporca/libnaucrates/include/naucrates/md/CMDFunctionGPDB.h
@@ -37,7 +37,7 @@ private:
 	CMemoryPool *m_mp;
 
 	// DXL for object
-	const CWStringDynamic *m_dxl_str;
+	const CWStringDynamic *m_dxl_str = nullptr;
 
 	// func id
 	IMDId *m_mdid;
@@ -101,11 +101,7 @@ public:
 	~CMDFunctionGPDB() override;
 
 	// accessors
-	const CWStringDynamic *
-	GetStrRepr() const override
-	{
-		return m_dxl_str;
-	}
+	const CWStringDynamic *GetStrRepr() override;
 
 	// function id
 	IMDId *MDId() const override;