--- a/src/backend/gporca/libnaucrates/include/naucrates/md/CMDTypeInt4GPDB.h
+++ b/src/backend/gporca/libnaucrates/include/naucrates/md/CMDTypeInt4GPDB.h
@@ -104,7 +104,7 @@ private:
 	IMDId *m_mdid_count;
 
 	// DXL for object
-	const CWStringDynamic *m_dxl_str;
+	const CWStringDynamic *m_dxl_str = nullptr;
 
 	// type name and type
 	static CWStringConst m_str;
@@ -127,11 +127,7 @@ public:
 								BOOL is_null) const override;
 
 	// accessors
-	const CWStringDynamic *
-	GetStrRepr() const override
-	{
-		return m_dxl_str;
-	}
+	const CWStringDynamic *GetStrRepr() override;
 
 	IMDId *MDId() const override;