--- a/src/backend/gporca/libnaucrates/include/naucrates/md/IMDCacheObject.h
+++ b/src/backend/gporca/libnaucrates/include/naucrates/md/IMDCacheObject.h
@@ -90,7 +90,7 @@ public:
 	virtual void Serialize(gpdxl::CXMLSerializer *) const = 0;
 
 	// DXL string representation of cache object
-	virtual const CWStringDynamic *GetStrRepr() const = 0;
+	virtual const CWStringDynamic *GetStrRepr() = 0;
 
 
 	// serialize the metadata id information as the attributes of an