… src/backend/gporca/libnaucrates/src/md/CMDFunctionGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDIndexGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDRelationCtasGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDRelationGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDScCmpGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDScalarOpGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeBoolGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeGenericGPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeInt2GPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeInt4GPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeInt8GPDB.cpp
… src/backend/gporca/libnaucrates/src/md/CMDTypeOidGPDB.cpp