With the GPDB7 partitioning work, Orca reads each individual partition's information instead of only the top level partition. This allows us to perform static elimination and is needed to construct the constraints for each child partition (and the column mapping for those constraints). However, this means that for a table with 1000 partitions and 1000 columns, we're doing much more processing than before. We're reading 1000 tables from the catalog, and 1000*1000 columns. Certain operations that were negligible before now take significant time and relcache space. The 3 optimizations in the commits here bring the optimization time down much more closely to the 6X numbers.

Together, these changes bring the optimization time for a simple select * from part_tbl with 1200 partitions and 200 int columns down from 6s to 1.1s for an empty relcache, and 560ms to 110ms with the tables in the relcache. In 6X, these numbers are closer to 200ms with an empty relcache, and 30ms within the relcache. There's still some other areas we can improve (reading the table names into a CmdName takes 250ms for example), but these are the major ones I found.

Fixes https://github.com/greenplum-db/gpdb/issues/14169