@chrishajas 2023-04-24
> For reference, can we show the improvement in optimization time using a query. For eg the query mentioned in Ref PR: #14169

I fixed up the individual commit messages to be more explicit about the example query I used. The explicit DDL is:
```
CREATE TABLE rank (
dist int,
key int,
a1 int,
...
a200)
DISTRIBUTED BY (dist)
PARTITION BY RANGE (key)
( START (1) END (1200) EVERY (1), 
  DEFAULT PARTITION extra ); 
explain select * from rank;
```