@chrishajas 2023-04-26
> Forward thinking: I think C++17/20 have much more library support for concurrency primitives (e.g. parallel for loops and async execution). If we can upgrade, it would be really interesting to see if we could leverage any of that in these hot paths.

I agree. I also think maybe passing around the map in certain cases instead of remaking it would remove some overhead. There's definitely room for more optimizations here.